requirements.txt
tensorflow==2.21.0
Pillow>=10.0.0
python-dotenv>=1.0.0


